Move link-renderer tests to it’s own file
